pyproject.toml
[project]
name = "fireworks-text-to-sql-takehome"
version = "0.1.0"
description = "FireworksAI Text-to-SQL Take-Home Assessment"
requires-python = ">=3.11"
dependencies = [
    "openai",
    "ipython",
    "pandas",
    "requests",
    "python-dotenv"
]
